Prevent race condition where `canopy_default_topic` can get erased
Previously we deleted the canopy_default_topic file before rewriting it
creating a time window where a sigint could delete it and break the
build going forward, so now we rewrite but don't erase first.

+++ cli/bulk/file_system_manager.js
@@ -6,6 +6,7 @@ let { DefaultTopic } = require('../shared/fs-helpers');
 class FileSystemManager {
   execute(fileSystemChange, logging) {
     fileSystemChange.fileDeletions.forEach(filePath => {
+      if (filePath === 'canopy_default_topic') return; // rewrite don't delete in case sigint
       fs.unlinkSync(filePath);
     });
